Update to Peripheral API v1.3.5 - Change int properties to floats

+++ src/storage/PrimitiveConfiguration.h
@@ -25,15 +25,15 @@ namespace JOYSTICK
 {
   struct TriggerProperties
   {
-    int center;
-    unsigned int range;
+    float center;
+    float range;
 
     TriggerProperties() { Reset(); }
 
     void Reset()
     {
-      center = 0;
-      range = 1;
+      center = 0.0f;
+      range = 1.0f;
     }
 
     bool operator==(const TriggerProperties& other) const
